Chapter10/requirement.txt
# Function dependencies, for example:
# package>=version
sqlalchemy
pymysql
google-cloud
google-cloud-pubsub
pandas
gcsfs
cloudstorage
google-cloud-storage